lattice based signature
